Remove uneeded cast to int that causes compiler error

+++ jpeg.d
@@ -1023,7 +1023,7 @@ private:
     }
     if (!rv)
     {
-      int capacity = JPGD_MAX(32768 - 256, (nSize + 2047) & ~2047);
+      size_t capacity = JPGD_MAX(32768 - 256, (nSize + 2047) & ~2047);
       mem_block *b = cast(mem_block*)jpgd_malloc(mem_block.sizeof + capacity);
       if (!b) { stop_decoding(JPGD_NOTENOUGHMEM); }
       b.m_pNext = m_pMem_blocks; m_pMem_blocks = b;
